requirements.txt
Flask
flask-cors
numpy
Pillow
scikit-learn
# chart.js 
# react-chartjs-2